created server and database connection
