+++ package.json
@@ -11,8 +11,10 @@
   "dependencies": {
     "cors": "2.8.5",
     "cross-env": "7.0.3",
+    "dotenv": "^16.4.5",
     "express": "4.17.1",
     "joi": "^17.13.1",
+    "mongoose": "^8.4.3",
     "morgan": "1.10.0",
     "uuid": "^10.0.0"
   },
